+++ sift_seq/encoder.py
